+++ sdk/package.json
@@ -52,20 +52,10 @@
     "typescript": "5.2.2"
   },
   "dependencies": {
-    "@aragon/osx-commons-configs": "^0.4.0",
-    "@aragon/osx-ethers": "^1.3.0-rc0.4",
-    "@aragon/osx-ethers-v1.0.0": "npm:@aragon/osx-ethers@1.2.1",
-    "@aragon/sdk-ipfs": "^1.1.0",
-    "@ethersproject/abstract-signer": "^5.7.0",
     "@ethersproject/bignumber": "^5.7.0",
-    "@ethersproject/constants": "^5.7.0",
     "@ethersproject/contracts": "^5.7.0",
     "@ethersproject/hash": "^5.7.0",
-    "@ethersproject/providers": "^5.7.2",
-    "@ethersproject/wallet": "^5.7.0",
-    "graphql": "^16.5.0",
-    "graphql-request": "^4.3.0",
-    "ipfs-http-client": "^51.0.0",
-    "yup": "^1.2.0"
+    "@ethersproject/providers": "^5.7.0",
+    "ipfs-http-client": "^51.0.0"
   }
 }
