+++ sdk/src/bitmap/utils.ts
@@ -4,6 +4,9 @@ import {
   InvalidBitPositionError,
 } from './errors';
 import {BigNumber} from 'ethers';
+// TODO:
+// - Revisit this in the future see if we can use flipBit
+// inside the boolArrayToBitmap function
 
 /**
  * Flips a specific bit in a `BigNumber` object and returns a new `BigNumber` object with the bit flipped.
