+++ sdk/src/ens/index.ts
@@ -1,1 +1,2 @@
 export * from './utils';
+export * from './errors';
