+++ sdk/src/ens/regex.ts
@@ -1,0 +1,2 @@
+export const ENS_REGEX = /^(?:[a-z0-9-]+\.)*[a-z0-9-]+\.eth$/;
+export const SUBDOMAIN_REGEX = /^[a-z0-9-]+$/;
