+++ sdk/src/ens/utils.ts
@@ -1,5 +1,5 @@
-import {InvalidEnsError, UnsupportedNetworkError} from '../errors';
-import {isEnsName} from '../validation';
+import {InvalidEnsError, UnsupportedNetworkError} from './errors';
+import {ENS_REGEX, SUBDOMAIN_REGEX} from './regex';
 import {getNetworkByAlias} from '@aragon/osx-commons-configs';
 import {Networkish} from '@ethersproject/networks';
 import {JsonRpcProvider, Provider} from '@ethersproject/providers';
@@ -34,4 +34,24 @@ export function resolveEnsName(
     provider = new JsonRpcProvider(aragonNetwork.url, providerOrNetwork);
   }
   return provider.resolveName(ensName);
+}
+/**
+ * Checks if the given string is a valid ENS name
+ *
+ * @export
+ * @param {string} name
+ * @return {*}  {boolean}
+ */
+export function isEnsName(name: string): boolean {
+  return ENS_REGEX.test(name);
+}
+/**
+ * Checks if the given string is a valid subdomain
+ *
+ * @export
+ * @param {string} name
+ * @return {*}  {boolean}
+ */
+export function isSubdomain(name: string): boolean {
+  return SUBDOMAIN_REGEX.test(name);
 }
