--- sdk/src/hex/constants.ts
@@ -1,6 +1,0 @@
-export const HEX_STRING_REGEX = /^(0x)?[0-9a-fA-F]*$/;
-export const OSX_EXTENDED_PROPOSAL_ID_REGEX =
-  /^(0x[A-Fa-f0-9]{40})_(0x[A-Fa-f0-9]{64})$/;
-export const OSX_COMPACT_PROPOSAL_ID_REGEX =
-  /^(0x[A-Fa-f0-9]{40})_(0x[A-Fa-f0-9]{1,64})$/;
-export const ADDRESS_REGEX = /^0x[A-Fa-f0-9]{40}$/;
