--- sdk/src/hex/index.ts
@@ -1,6 +1,0 @@
-export * from './proposal';
-export * from './errors';
-export * from './address';
-export * from './bytes';
-export * from './constants';
-export * from './strings';
