+++ sdk/src/index.ts
@@ -1,23 +1,11 @@
-export * from './constants';
-export * from './types';
-export * from './validation';
-export * from './multiuri';
-export * from './ipfs';
-export * from './metadata';
-export * from './interfaces';
-export * from './promises';
 export * from './errors';
-
-// SDK
-
-// From OSX
 export * from './bitmap';
 export * from './events';
-export * from './math';
 export * from './permission';
-export * from './proposal';
 export * from './time';
-
 export * from './ens';
-export * from './plugin';
-export * from './hex';
+export * from './proposal';
+export * from './ratio';
+export * from './metadata';
+export * from './ipfs';
+export * from './introspection';
