+++ sdk/src/ipfs/utils.ts
@@ -1,0 +1,37 @@
+import {IPFS_CID_REGEX, IPFS_URI_REGEX} from './regex';
+import IPFS from 'ipfs-http-client';
+
+// TODO Revisit
+export async function uploadToIPFS(content: string): Promise<string> {
+  const client = IPFS.create({
+    url: 'https://prod.ipfs.aragon.network/api/v0',
+    headers: {
+      'X-API-KEY': 'b477RhECf8s8sdM7XrkLBs2wHc4kCMwpbcFC55Kt',
+    },
+  });
+
+  const res = await client.add(content);
+  await client.pin.add(res.cid);
+  return res.cid.toString();
+}
+
+/**
+ * Checks if the given string is a valid IPFS CID
+ *
+ * @export
+ * @param {string} cid
+ * @return {*}  {string}
+ */
+export function isIpfsCid(cid: string): boolean {
+  return IPFS_CID_REGEX.test(cid);
+}
+/**
+ * Checks if the given string is a valid IPFS URI
+ *
+ * @export
+ * @param {string} cid
+ * @return {*}  {boolean}
+ */
+export function isIpfsUri(cid: string): boolean {
+  return IPFS_URI_REGEX.test(cid);
+}
