+++ sdk/src/metadata/types.ts
@@ -1,0 +1,7 @@
+export type MetadataAbiInput = {
+  name: string;
+  type: string;
+  internalType: string;
+  description: string;
+  components?: MetadataAbiInput[];
+};
