+++ sdk/src/metadata/utils.ts
