--- sdk/src/permission.ts
@@ -1,63 +1,0 @@
-import {ADDRESS} from './constants';
-import {id} from '@ethersproject/hash';
-
-export enum Operation {
-  Grant = 0,
-  Revoke = 1,
-  GrantWithCondition = 2,
-}
-
-export const PERMISSION_MANAGER_FLAGS = {
-  UNSET_FLAG: ADDRESS.ZERO,
-  ALLOW_FLAG: ADDRESS.TWO,
-  ANY_ADDR: ADDRESS.LAST,
-  NO_CONDITION: ADDRESS.ZERO,
-};
-
-export const DAO_PERMISSIONS = {
-  ROOT_PERMISSION_ID: id('ROOT_PERMISSION'),
-  EXECUTE_PERMISSION_ID: id('EXECUTE_PERMISSION'),
-  UPGRADE_DAO_PERMISSION_ID: id('UPGRADE_DAO_PERMISSION'),
-  SET_SIGNATURE_VALIDATOR_PERMISSION_ID: id(
-    'SET_SIGNATURE_VALIDATOR_PERMISSION'
-  ),
-  SET_TRUSTED_FORWARDER_PERMISSION_ID: id('SET_TRUSTED_FORWARDER_PERMISSION'),
-  SET_METADATA_PERMISSION_ID: id('SET_METADATA_PERMISSION'),
-  REGISTER_STANDARD_CALLBACK_PERMISSION_ID: id(
-    'REGISTER_STANDARD_CALLBACK_PERMISSION'
-  ),
-  VALIDATE_SIGNATURE_PERMISSION_ID: id('VALIDATE_SIGNATURE_PERMISSION'),
-};
-
-export const PLUGIN_UUPS_UPGRADEABLE_PERMISSIONS = {
-  UPGRADE_PLUGIN_PERMISSION_ID: id('UPGRADE_PLUGIN_PERMISSION'),
-};
-
-export const PLUGIN_REPO_PERMISSIONS = {
-  ROOT_PERMISSION_ID: id('ROOT_PERMISSION'),
-  MAINTAINER_PERMISSION_ID: id('MAINTAINER_PERMISSION'),
-  UPGRADE_REPO_PERMISSION_ID: id('UPGRADE_REPO_PERMISSION'),
-};
-
-export const PLUGIN_SETUP_PROCESSOR_PERMISSIONS = {
-  APPLY_INSTALLATION_PERMISSION_ID: id('APPLY_INSTALLATION_PERMISSION'),
-  APPLY_UPDATE_PERMISSION_ID: id('APPLY_UPDATE_PERMISSION'),
-  APPLY_UNINSTALLATION_PERMISSION_ID: id('APPLY_UNINSTALLATION_PERMISSION'),
-};
-
-export const ENS_REGISTRAR_PERMISSIONS = {
-  UPGRADE_REGISTRAR_PERMISSION_ID: id('UPGRADE_REGISTRAR_PERMISSION'),
-  REGISTER_ENS_SUBDOMAIN_PERMISSION_ID: id('REGISTER_ENS_SUBDOMAIN_PERMISSION'),
-};
-
-export const PLUGIN_REGISTRY_PERMISSIONS = {
-  UPGRADE_REGISTRY_PERMISSION_ID: id('UPGRADE_REGISTRY_PERMISSION'),
-  REGISTER_PLUGIN_REPO_PERMISSION_ID: id('REGISTER_PLUGIN_REPO_PERMISSION'),
-  ENS_REGISTRAR_PERMISSIONS,
-};
-
-export const DAO_REGISTRY_PERMISSIONS = {
-  UPGRADE_REGISTRY_PERMISSION_ID: id('UPGRADE_REGISTRY_PERMISSION'),
-  REGISTER_DAO_PERMISSION_ID: id('REGISTER_DAO_PERMISSION'),
-  ENS_REGISTRAR_PERMISSIONS,
-};
