--- sdk/src/plugin/index.ts
@@ -1,3 +1,0 @@
-export * from './utils';
-export * from './errors';
-export * from './proposal';
