+++ sdk/src/ratio/errors.ts
