+++ sdk/test/constants.ts
@@ -1,20 +1,6 @@
 import {MetadataAbiInput} from '../src';
 
-export const TEST_WALLET =
-  '0xdf57089febbacf7ba0bc227dafbffa9fc08a93fdc68e1e42411a14efcf23656e';
-export const TEST_WALLET_ADDRESS = '0x8626f6940E2eb28930eFb4CeF49B2d1F2C9C1199';
-export const ADDRESS_ZERO = '0x0000000000000000000000000000000000000000';
 export const ADDRESS_ONE = '0x0000000000000000000000000000000000000001';
-
-export const web3endpoints = {
-  working: ['https://cloudflare-eth.com/'],
-  failing: ['https://bad-url-gateway.io/'],
-};
-
-export const DEFAULT_IPFS_ENDPOINTS = [
-  'https://prod.ipfs.aragon.network/api/v0/',
-  'https://test.ipfs.aragon.network/api/v0/',
-];
 
 export const TEST_ABI: MetadataAbiInput[] = [
   {
@@ -109,10 +95,6 @@ export const TEST_ABI: MetadataAbiInput[] = [
   },
 ];
 
-export const TEST_ADDRESS = '0x0000000000000000000000000000000000000001';
-export const TEST_INVALID_ADDRESS =
-  '0x000000000000000000000000000000000000000P';
-
 export const TEST_ENS_NAME = 'subdomain.test.eth';
 export const TEST_INVALID_ENS_NAME = 'test.invalid';
 
@@ -127,28 +109,6 @@ export const TEST_INVALID_IPFS_CID = '1nv4l1d_c1d';
 export const TEST_INVALID_IPFS_URI = `ipfs://${TEST_INVALID_IPFS_CID}`;
 
 export const TEST_HTTP_URI = 'https://test.com';
-export const TEST_INVALID_HTTP_URI = 'https://te?st.com-invalid';
 
 export const TEST_SUBDOMAIN = 'test';
 export const TEST_INVALID_SUBDOMAIN = 'test.invalid';
-export const ERC165_ABI = [
-  {
-    inputs: [
-      {
-        internalType: 'bytes4',
-        name: 'interfaceId',
-        type: 'bytes4',
-      },
-    ],
-    name: 'supportsInterface',
-    outputs: [
-      {
-        internalType: 'bool',
-        name: '',
-        type: 'bool',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-];
